QAI-10345: Fix negative std-dev in sfnt-pingpong
The std-dev result for runs of sfnt-pingpong are sometimes negative.
This is caused, in the variance calculation, by 'sumsq' being a
signed 64-bit int, so when diff * diff is calculated this may
overflow in positive 64-bit space. Changing the type of 'sumsq'
and 'variance' to double prevents this (though may sacrifice
precision).

+++ src/sfnettest.h
@@ -304,7 +304,7 @@ extern void sfnt_iarray_mean_and_limits_int64(const int64_t* start, const int64_
                                int64_t* mean_out, int64_t* min_out, int64_t* max_out);
 
 extern void sfnt_iarray_variance_int64(const int64_t* start, const int64_t* end,
-                        int64_t mean, int64_t* variance_out);
+                        int64_t mean, double* variance_out);
 
 /**********************************************************************
  * File / muxer convenience functions.
